update: refactor the ts fils to js
